Split EventListener into different files
