+++ lib/event_listener.rb
@@ -12,59 +12,4 @@ module Ecco
       @type = event.get_header.get_event_type
     end
   end
-
-  class SaveEventListener < EventListener
-    SAVE_EVENTS = [
-      EventType::QUERY,
-      EventType::ROTATE,
-      EventType::WRITE_ROWS,
-      EventType::UPDATE_ROWS,
-      EventType::DELETE_ROWS,
-    ]
-
-    def on_event(event)
-      super(event)
-
-      case @type
-      when *SAVE_EVENTS
-        filename = @client.get_binlog_filename
-        position = @client.get_binlog_position
-
-        @client.on_save_position_callback.call(filename, position)
-      end
-    end
-  end
-
-  class RowEventListener < EventListener
-    RowEvent = Struct.new(:type, :table_id, :database, :table, :rows)
-
-    ROW_EVENTS = [
-      EventType::WRITE_ROWS,
-      EventType::UPDATE_ROWS,
-      EventType::DELETE_ROWS,
-    ]
-
-    def on_event(event)
-      super(event)
-
-      case @type
-      when EventType::TABLE_MAP
-        @table_map_event = event
-      when *ROW_EVENTS
-        row_event          = RowEvent.new
-        row_event.type     = @type.to_s
-        row_event.table_id = @data.get_table_id
-        row_event.rows     = @data.rows
-
-        if @table_map_event
-          event_data = @table_map_event.get_data
-
-          row_event.database = event_data.get_database
-          row_event.table    = event_data.get_table
-        end
-
-        @client.on_row_event_callback.call(row_event)
-      end
-    end
-  end
 end
